Update windy_gridworld_env.py
The problem happens when you create the environment and you try to render it without reseting first. 
If you don't specify the message of the error here, the error raised is: "AttributeError: 'WindyGridWorldEnv' object has no attribute 'observation'" which is not intuitive for what the user should do. 
But specifying that the problem states in first resetting the environment it is much better.